lab2/app/requirements.txt
streamlit==1.45.1
pandas==2.2.3
psycopg2-binary==2.9.10
plotly==6.1.2
sqlalchemy==2.0.19
pydeck==0.8.0
